Allow changing the schema in a stacked DB when glean.schema_id is specified
Summary:
If the client specifies `glean.schema_id`, then we know which schema they want. In a stacked DB we should therefore allow this under the same conditions as `--update-schema-for-stacked`, namely that the stacked schema should agree with the base schema for existing facts.

Also updated the docs for `glean create` with this info.

Reviewed By: lazamar

Differential Revision: D44710637

fbshipit-source-id: 95af83e82469a138bd8b1274b1c9ee703d3dfae9