Support scope search for C++ namespace symbols
Summary:
We haven't been able to search for namespace definitions previously, using terms like "apache::thrift" in Glass.

Where we have a scope term, and it resolves to a namespace fact, check if the pair of name and scope represent a new NamespaceQName. Iterate across the namespace decls to find defns and off we go. If the scope term is `global_ = {}` we can use that as a nothing marker for namespaces at the root (e.g. `facebook` or `folly`).

Namespace declarations are tricky: we don't have a single definining entity, instead thousands of instances.
We should be able to build a reasonable report of the API by aggregating definitions now however. Let's start with search.

Previously, none of these namespace symbols would be found, only when there was a non-namespace symbol suffix on the query.

Some useful examples that work now:
- `apache::thrift`
- `apache::thrift::compiler::cpp2`
- `facebook::adindexer`
- `facebook::petra`
- `facebook::nodeapi`
- `facebook::tupperware::api`
- `folly`
- `facebook::stl`
- `facebook::koski`
- `apache::thrift::transport`
- `facebook::infrasec::authorization`

## footnotes

Uncovered a codex/doxygen bug, where codex believes "namespace folly::dynamic" is a thing. It isn't.
https://pxl.cl/2CgVD

Reviewed By: simonmar

Differential Revision: D44905761

fbshipit-source-id: 4966f81b8cba4e4c62b9ece505fd7c3da2182217