Handle Range-based languages
Summary:
The Clang indexer produces `Range` ranges instead of `ByteSpan` ranges. The Digest deriver had a TODO for these, since they are only used in Clang languages.

This Diff addresses that TODO. Range ranges are different in two ways:

1. They are line/col instead of index/length
2. They contain a file fact so they can reference a different file than the one containing the entity

The first point is addressed by converting to index/length using `src.FileLines` facts retrieved from the DB. The conversion is efficient only when batched for all the ranges in a file.

The second point needs careful rearranging of the text slicing logic in the deriver, since the slices are efficient only when batched for all the spans in a file.

Reviewed By: phlalx

Differential Revision: D45046659

fbshipit-source-id: a0424d09ce17fa540b406dc2cad2f18f43039b38