basic schema with methods support
Summary:
Initial schema supporting Methods indexing for Kotlin glean.
The schema was added as "kotlin.alpha" to add flexibility in incompatible changes during development using versions bump and to release "kotlin.1" when everything will be ready.
Common javakotlin schema is going to be a synchronisation place between 2 languages. declarations in languages are going to be linked via QNames.

Reviewed By: nhawkes

Differential Revision: D44138860

fbshipit-source-id: f69ab72a65c1721238bbe9691f64c9ae7f1b35d7