Add a "hide uninteresting" version of search related parents via `extends`
Summary:
We plan to support hiding some parts of the inheritance DAG, ideally in Glean.
For Hack, there's an annotation that let's us know to drop the container.

This implements the plumbing to get that flag set.

The implementation of Parent via Extends and the concise version are the same, just one with an additional filter.

Reviewed By: simonmar

Differential Revision: D45214220

fbshipit-source-id: a79edc39a31c08d966a66fb4e95dbad22c6df572