Deterministically sort targets with the same byte spans.
Summary:
The predicate `FileXRefs` maps a variable list of byte spans `[src.ByteSpans]` in `FileXRefMap` to a list of external xref targets `[XRefTarget]`.

There are scenarios under which the same byte spans will refer to many `XRefTarget`.

One scenario is xrefs formed under macros:
```
#define FOO foo(bar())
```
A use of this macro will end up xrefs from the use-site to `foo` and `bar`.

Another scenario is xrefs formed by an overloaded expression.

```
void foo(int) {}
void foo(double) {}

template <typename T>
void bar(T t) {
  foo(t);
}
```

In this case, we end up with xrefs from the `foo` call to both overloads.

Now, these target sets can be formed in various ways depending on what's included at the time of processing a header that this code is a part of. The targets will be seen in different order, they'll be different sets altogether, etc.

This diff deterministically sorts the targets such that we'll end up with the same list if we have the same set of targets. This in turn reduces the # of instances of `FileXRefs` since we get much more de-duplication.

Reviewed By: simonmar

Differential Revision: D44904750

fbshipit-source-id: 6d386e3de3e6ae182fd74c62c04d1aa8a6365e6d