De-dup src.File facts in the indexer since its easy
Summary:
When converting SCIP to Glean JSON, we were generating a unique src.File fact for every definition and against for ever set of occurences. Since we have a two pass algo, we can just record the initial src.Fact id and reuse it in the second pass.

Saves 1% on disk in the text/json format:

```
$ du -hs ~/index.scip
38M	/home/dons/index.scip
```
becomes:
```
$ du -hs ~/scip.json
95M	/home/dons/scip.json
```

was
```
$ du -hs ~/scip-old.json
96M	/home/dons/scip-old.json
```

Reviewed By: simonmar

Differential Revision: D45804595

fbshipit-source-id: 741a3930257e321eef2b3710ceda817c77210114