Direct scip.angle encoding of scip.proto + indexer
Summary:
Adds scip.angle, a direct translation of scip.proto, where the scip.Symbol is the shared key between predicates. And rewrites the SCIP to Angle converter to target scip.angle instead of lsif.angle.

scip.angle directly captures the precise rules of the scip.proto encoding, rather than trying to translate to lsif-like encodings on the fly. A lot simpler and more precise.

The reason to do this is to permit sharding / concurrent indexing, where xrefs from one scip file refer to definitions in another only by the `symbol` field.

This moves deriving xref <-> def relationships to the Glean stage, rather than directly in indexing. Only after all scip files are written do we have the complete set of relationships.

Next:
- define the entity type
- derive xref <-> defn relationships
- add codemarkup layer
- add symbol lookup layer
- glass

Much of this is similar to LSIF but drops a bunch of the baggage.

## notes on linting

The scip file isn't detected as a binary file, so hits a utf8 warning.
We can do a fun thing to suppress it:
```
sed -i '3s;rust-analyzer;generated   ;' glean/lang/scip/tests/cases/rust-analyzer/index.scip
```

i.e. change the binary data for the indexer name to "generated  "  (instead of "rust-analyzer" , and then the linters are kosher.

> Bypassing the utf8 check as its not able to detect index.scip is a binary file (as there are no null bytes in the protobuf data). It is a binary file so skipping land-block as per: https://www.internalfb.com/intern/qa/269361/binary-file-is-failing-invalid-utf8

Reviewed By: pepeiborra

Differential Revision: D45839757

fbshipit-source-id: 0776d8446c0c5c5bf6c31885e77860b53713901b