strict error handling
Summary:
Instead of returning an empty response and logging the problem, throw a client error explaining the problem. This is an opt-in change of behaviour via a new `RequestOptions` flag.

I have updated the Glass CLI to always run in strict mode, which is more useful than the current behaviour

Reviewed By: donsbot

Differential Revision: D45733098

fbshipit-source-id: f38e326bbc2bc9aa0c2d4b40124deeae6723d50d