Fix relative CDBs in the indexer
Summary:
We want to produce relative CDBs (where `file` is relative to the root
of the repo and `directory = "."`, for example), because those are
cacheable by Buck2 and work in RE.

But: relative CDBs are not actually supported by
ClangTool. `ClangTool.run()` converts the input source file name to an
absolute path before looking it up in the CDB, so if the CDB contains
relative paths this lookup will always fail.

Fix: transform the CDB when we load it to replace `directory="."` with
`directory="<cwd>"`, and then everything works.

Reviewed By: donsbot

Differential Revision: D45986157

fbshipit-source-id: 44a97301225b93bb0706a84d8f59023e3507c4a0