Fix find all references
Summary:
Got report from user:

https://www.internalfb.com/code/fbsource/[99847db6d501843d35fda186e997172ca23dbdc9]/fbcode/mcrouter/lib/network/AsyncMcServer.cpp?lines=567

Click-to-defs to
https://www.internalfb.com/code/fbsource/fbcode/folly/io/async/AsyncServerSocket.cpp

But find-refs doesn't work
 ---
My best guess is the find refs is going on the .h file, but the ref is to the cpp file.

Either way doing a proper union of all the entities rather than just the first one seems to work. As a bonus this'll fix find-refs for multiple dbs

Reviewed By: pepeiborra

Differential Revision: D45946122

fbshipit-source-id: 361fedd1b5fae67732975a87fca736e6968e5da8