Remove an unused field `USRSymbolDefinition::nameRange`.
Summary:
This diff tries to remove an unused field `nameRange` from `USRSymbolDefinition`. This is in preparation to a migration from `nameRange` of type `Range` to `nameLocation` of type `LocationRange`.

In other places such as `ReferenceRangeSymbolX`, `DefinitionSymbolX`, and `SymbolX`, we'll be migrating the clients to use a new field of type `LocationRange`.

But this `nameRange` is unused, we'll just remove it for now and add a `LocationRange` later if we end up wanting it.

Reviewed By: donsbot

Differential Revision: D46058268

fbshipit-source-id: ae909696f5d8a8be03eef73902d78b149294fb78