serialize all method declarations
Summary:
We are going to start integrating the indexer into the infra with MethodDeclaration as the root object and then we are going to expand it to class/files/packages.
Before this diff for simplicity we were serializing only last MethodDeclaration, the diff adds support for multiple methods and aggregates methods into JSON array.
It is a temporary solution, because in the future Methods are not going to be root objects and we are going to have single root object per module/index.

Reviewed By: kathryngray

Differential Revision: D46021568

fbshipit-source-id: 5f8b607dcc7d72460e0462a6f17cd93c1f0a2bb8