write index to buck output
Summary:
We should deligate output folder management to Buck instead of creating it manually. It simplifies folders management and potentially can allow to cache output.
Kotlin Step in Buck puts everything from output folder to a final jar, it allows us to have the same approach in Kotlin as we currently have in Java (where index is added to jar)

Reviewed By: kathryngray

Differential Revision: D46020088

fbshipit-source-id: 86652e4097b209451b026a1ef621b35ea413f560