Introduce new types of errors
Summary:
Problem: src.IndexFailure has only one reason 'compileError'. There are places where we loose file for other reasons. So we need this enum to be able to properly incorporate that

Propose to use 'BuildSystemError' to indicate that we cannot determine a build system or missing some information to even run a build command for this file.

Reviewed By: simonmar

Differential Revision: D46179377

fbshipit-source-id: 5d303a5f6e1966d73e7f78143a5b17b4bbb1505c