Remove `spellingRange` and leave a comment for `srcText`.
Summary:
Computing the spelling range given an expansion range is a rather difficult task without a clear answer. We currently have a naive implementation of spelling range which is only used to retrieve the source text of attributes.

This diff removes the naive `spellingRange` function and moves it into `srcText`, with a large comment for `srcText`. Perhaps I'll move it all the way to the attribute indexing part, but `srcText` seems like a rare enough thing for us to want to do, and a comment seems sufficient (at least better than before) for now.

Reviewed By: simonmar

Differential Revision: D46172765

fbshipit-source-id: 8a1edebbb3603fd230ca17dcea0f3dfb4f411322