Change the `span` field of `codemarkup.Location` to `destination`.
Summary:
The `cxx1.DeclarationNameSpan` predicate maps a declaration to the location of its name as a (`file`, `span`) pair. Currently, it's safe to assume that the name location is in the same file as the declaration.

Specifically, in the case of a declaration within a macro definition:

```
// a.h
#define DEFINE_FOO void foo() {}

// a.cpp
DEFINE_FOO
```

The range of the `foo` declaration as well as the name of `foo` is the range of the use of `DEFINE_FOO` in `a.cpp`.

However, we'll be changing this such that the range of `foo` declaration is the range of the use of `DEFINE_FOO` in `a.cpp`, but the location of the name of `foo` is the range of `foo` in the definition of `DEFINE_FOO` in `a.h`.

This diff changes the `span` field in `codemarkup.Location` to `destination`. We still report the span for the existing Glass Thrift fields, but we filter out the `destination`s that are not in the same file or not a subrange of the declaration, so that the clients of those fields can still safely assume that the spans are in the same file as the declaration.

Reviewed By: simonmar

Differential Revision: D46058369

fbshipit-source-id: b788a40cf35a658f93934630d8d5a911a875816a