use faster query for call hierarchy in C++
Summary:
This uses the new `codemarkup.EntitySource` predicate which is much more efficient than `codemarkup.ReferencingEntities`. It is only available in `fbsource` and `www.hack`, but here we only use it on the former because Glass doesn't currently use `www.hack` for anything.

While `codemarkup.EntitySource` doesn't provide call site ranges, this is not needed for call hierarchy.

Reviewed By: simonmar

Differential Revision: D46391628

fbshipit-source-id: 1933a68b578b96443cda00306ab78421bfdea716