Only search for defs if the decl is not already a def.
Summary:
It seems like we have weird build scenarios where `fbcode` and `xplat` versions of `folly` exist together.

This makes it such that we output `cxx1.Same` facts which later get derived into `cxx1.DeclFamily` and related predicates.

During the decl -> def search, we make use of this to find the `code.cxx.Definition` entity within a `cxx1.DeclFamily`. In our case, we can have multiple definitions of say `fbcode/folly/dynamic` and `xplat/folly/dynamic`.

> The fact that we derive the full `DeclFamily` after all `Same` facts have been generated, actually seems somewhat bad... since the entire DB can be "polluted" by some weird thing in a corner.

If we have an xref to `fbcode/folly/dynamic`, following through the logic of `CxxDeclToDefXRefTargetLocation` we find `xplat/folly/dynamic`, and filter ourselves out. This logic I believe is due to it usually being the case that we perform `code.cxx.DeclToDef`, get a single result back and we filter that out since we're already a definition.

This diff basically performs the "are we already a definition" check earlier, so that if we have the definition of `fbcode/folly/dynamic`, we just keep it.

Reviewed By: donsbot

Differential Revision: D46385028

fbshipit-source-id: 1684842fd631662cbe75b447a982903c21d25ab5