make buck.FailureSource predicate faster
Summary:
I hope this is the allowed change :)

This is the current definition:
predicate buck.FailureSources.4 :
    { error : buck.IndexFailureX.4
    , source : src.File.1
    }

Given this, I would expect that if I have a buck.indexFailureX.4 fact, I can efficiently find its source. Like this:
```
MyError = …… ;
buck.FailureSources { MyError, _ };
```
This goes in line with the idea that in Glean prefix searches are efficient.
But that is actually not the case for this predicate.
In its implementation, this will trigger a full scan of buck.FileToTarget.4, because of the suffix search because this query is inefficient:
  Error = buck.IndexFailureX { source = {locator = Locator}};
  buck.FileToTarget { source = Src,locator = Locator}

For each locator found in the first line it looks through all facts of buck.FileToTarget to find one with the locator in a suffix position. This means that a prefix search of this derived predicate is not efficient.

However, if I have a src.File I can use it to efficiently find a buck.FileToTarget.4 fact. Without any full scan, only using prefix searches.

This means that we would expect the src.File field to come first in the derived predicate record.

Reviewed By: lazamar

Differential Revision: D46556363

fbshipit-source-id: 625098582e3a8b46d307d4232e45f11d16eef522