Start building the Java side of Java/Kotlin glean schema
Summary:
Creating a new alpha variant of Java schema that interops with Kotlin.

This is not meant to directly build/evolve the existing Java schema due to the interop and thus would be a breaking change, we will have a new db from the existing Java db so they will not need to co-exist.

There is no matching indexer support yet.

Reviewed By: nhawkes

Differential Revision: D45608326

fbshipit-source-id: 4dff5a5c677b47bd4e01c3998fc64e03e88160a9