Allow references to remote facts on SendAndRebaseQueue
Summary:
To allow remote references we need to turn off type checking on the client side as we cannot check remote facts.
If there is a type error it will be thrown on the server side and the client will not be able to recover.
This outcome is not too different from what would happen with local type checking as an indexer that produces a type error doesn't learn to not produce those during the course of an indexing run.

Given this feature causes local fact ids to be large (9 bytes), it is behind a boolean flag `sendAndRebaseQueueAllowRemoteReferences`.

Reviewed By: simonmar

Differential Revision: D45903845

fbshipit-source-id: d8b60c69dd5884c5c706b78e68d40f2c6f1f044c