Make a branch of Java indexer for interop with kotlin
Summary:
This branch of Java indexer uses the java_alpha and javakotlin_alpha schemas.

It is a straight port of moving the indexer to these schemas and attempting to get the indexer to compile with the changes. No attempt was made to address any other issues to eliminate churn.

It is known that the tests are not running. The predicates will not currently be consistent/correct due to parameters needing a container and method definitions needing a parameter. This will be fixed in future iterations from this point, and then tests can be build after this is fixed.

There may be other missing predicate fields. This was not addressed in this port since the tests are not building and the aim was to get the indexer minimally compiling again.

Reviewed By: rybalkinsd

Differential Revision: D45982691

fbshipit-source-id: 3305ca65ef87fa99792aac405ebf8306f003522f