SendAndRebaseQueue: better stats
Summary:
Measure throughput of writes and throughput to the server, as well as
cache hit rate. I'm doing all this using the same stat framework that
the write server uses, since there's a lot in common between the two
cases.

It turns out that the cache is more than 99% effective at de-duping
facts before they get sent to the server.

Reviewed By: lazamar

Differential Revision: D46649483

fbshipit-source-id: 00dd98e6b2d60989afdf4f058c00cc1e3ad06f6e