Fix perf bug in subst.firstFreeId()
Summary:
I noticed that `glean write --use-local-cache` is sometimes taking a
VERY long time with JSON files for `www.hack.light`. `perf` pointed to
`Substitution::rebaseIntervals` as the culprit, and it turns out that
we call this function many times after receiving a substitution, but
each call invokes `firstFreeId()` which is O(n) in the size of the
substitution.

Fix: memoize `Substitution::firstFreeId()`

Reviewed By: lazamar

Differential Revision: D46721930

fbshipit-source-id: bef2c053655a767ca3054295083a5fb7f64b45c5