Support for selectively expanding results by predicate
Summary:
Sometimes for efficiency reasons you want to return results with
some but not all of the nested facts expanded. This diff adds support
for choosing which nested facts to fetch by predicate. The query
engine already had support for this (it's used by the legacy JSON
query API) but it wasn't exposed for Angle.

Added support in the DSL, so you can say

```
   expanding Some.Predicate $ query $ ...
```

and in the shell, you can do

```
:expand Some.Predicate
```

Reviewed By: mpark, lazamar

Differential Revision: D46761102

fbshipit-source-id: 71120a73ece9faf68da2a0f191f4cb308307ce5f