don't remove .out files agressively in regression tests
Summary:
When the launcher is called with `--replace` on a failing, or interrupted, indexer all `.out/.perf` files are deleted before regeneration.

This diff changes the behavior slightly. It doesn't remove .out files before regeneration, but only those that weren't regenerated.

Initial motivation:

When a test cases dir doesn't contain any .out files, it's not considered by the test launcher "discovery" phase. Combined with the .out disappearing, this be confusing.

Reviewed By: malanka

Differential Revision: D46877395

fbshipit-source-id: 283dc238ccc7bd46e72bd0bc23b9c6183846d92d