Fix size estimate for JSON write batches
Summary:
The x2 was wrong, after a little digging it looks like it was leftover
from when the batch was represented by a `Text` (aka UTF-16), and when
we switch to `ByteString` I forgot to change it: D18018132

The size estimate doesn't take into account the data structure
overhead either, so it could be off. Let's keep an eye on the queue
size and memory usage when this is landed and see if we need further
tweaks.

Reviewed By: malanka

Differential Revision: D46969238

fbshipit-source-id: 7b33f5f12f18e13f246df52f22f8a2a6c8ee4645