Return query results in the order they were produced
Summary:
mcypark discovered (D46966475) that query results are out of order
sometimes. When querying for `allFacts` of some predicate we expect to
see results in key order, and in fact the deriving pass was relying on
that.

The ordering gets lost because we're returning the results in a `map`,
and then turning the map into a list on the client side, which will
sort the results by `Id`, losing the original key ordering.

This doesn't provide a strict guarantee that the results will be
ordered, but for simple queries like `allFacts` for a stored predicate
I think we can assume that results will be ordered, and it's really
useful to be able to do that sometimes.

Reviewed By: mpark, malanka

Differential Revision: D46972649

fbshipit-source-id: 08d6584c93d38a980a86617d29a274daa5fe12e9