Fix digests deriver crashes
Summary:
The regexp hack we use to rename identifiers in the digest deriver can break for operators and for C++ identifiers.
This Diff handles the exception and returns the un-renamed thing.

To keep things simple, no logging is performed for now. I'll leave it to jmaessen to decide what logging is needed

Reviewed By: simonmar

Differential Revision: D46796217

fbshipit-source-id: ff1e7379fe25af61ed9202fcfe15ace68c75d040