A bit of paranoia
Summary:
If an `OwnershipUnitCopy` accidentally gets copied then things could
break in interesting ways because we have a `folly::Range` pointing to
the contents. So let's avoid that problem by making it safe to copy (I
tried the alternative of making it uncopyable but couldn't get the
compiler to accept it).

Reviewed By: lazamar

Differential Revision: D47096870

fbshipit-source-id: 4b52a5c39a6bc2a8c319f66f1c79fa72e73d52ae