Update the C++ indexer to fill in the new schema for better macro support.
Summary:
This diff continues on the macro work. We use the `fullSrcRange` function to get the "file"/"expansion"/"spelling" spans and track them separately inside `FileData` if the span is in the same file, or emit the `SpellingXRef` facts if the span is in a different file.

We have some extra code that combines all of them to "spans" to form the old schema during migration.

Reviewed By: simonmar

Differential Revision: D47145151

fbshipit-source-id: 7034d0a0c33fda97415da8db178f446d9f0aaa28