symbol index using name ranges
Summary:
The symbol index for a file is a mapping line->[symbol]. Codehub and other clients use the symbol index to O(1)-ish find the symbol at a (file, column, line) position.

The assumption that Codehub and other clients make is that the line used to index into the mapping is the line where the symbol is located. However, currently it is the starting line of the whole entity. These two notions coincide when the entity identifier is in the starting line, but this is not always the case.

In order to make the assumption valid, let's fix the mapping so that it uses the identifier line when available, falling back to the entity starting line when not available.

Reviewed By: donsbot

Differential Revision: D47055724

fbshipit-source-id: 8790478465dff8e3cb3f13106027051cd260a322