Allow the expansion of recursive predicates
Summary:
Calculates a particular transitive closure in the tests. Are done?

To get started we need to get recursive queries to compile at all.
The first step is allowing the flattening of recursive derivations (that's what this diff does), then we can start iterating on improving the code that is generated.

Reviewed By: simonmar

Differential Revision: D47191271

fbshipit-source-id: d2e091af4e12fbae5a18db636d3573b01891b0cc