Fix correctness bug in parallel de-duplication
Summary:
Parallel de-duplication takes a snapshot of the underlying DB to
de-duplicate against, to protect against seeing facts that might
have overlapping fact IDs with the batch being de-duplicated.

However, the snapshot of the DB was non-deterministic, because it
included facts that were being conccurrently written to the DB. This
meant that looking up a key in the snapshot was non-deterministic,
which could lead to a situation where two X->Y facts have the same X
but different Y, which is reported as `invalid fact redefinition`.

Reviewed By: donsbot, bochko, lazamar

Differential Revision: D47210623

fbshipit-source-id: cf923abbfebdac40c4690b15e17eae0b1b92df6a