Strip "fbsource' prefix on paths
Summary:
src.File facts in the Java and Kotlin dbs need to be anchored relative to fbsource/*

Do a two pass regex anchored at the start of the path to remove leading absolute path components up to /fbsource/, and then any noise.

I think a better approach would be to pass "--indexer-root" through the indexer but Im' not sure if that's possible with the compiler plugin model?

Reviewed By: kathryngray

Differential Revision: D47181309

fbshipit-source-id: 2b0ac01d24946f1ba2ca2cb52d35b19f2a329dd8