Test cpp diff sketching
Summary:
Right now C++ diff sketching isn't tested in our buck flow, and that's going to make it much harder to find and fix bugs such as T154688348.  Run the C++ sketcher & deriver during testing by invoking it as a library.  This is requiring a bunch of fixes from pepeiborra to how include handling is treated.

This fixes problems with file inclusion during indexing that were causing indexing to die with P770835297.

Reviewed By: pepeiborra

Differential Revision: D46649980

fbshipit-source-id: 69fecbdcc0961ff0b5fd440365199542d1681395