Log whether queries perform a full scan
Summary:
This diff adds a `full_scan` property to `UserQueryStats` which lists which predicates the compiled code would perform a full table scan over.

It would be nice to be able to quickly spot queries that can easily be optimised. These often are full table scans through suffix searches.
If we log which queries make full scans we could find the problematic and easily optimisable ones in Scuba by looking at the intersection of queries that take long and perform full cans.

The pruning optimisation becomes useful here as well as we won't count branches that would otherwise perform a full scan but which won't be evaluated because of the db's content.

This is obviously not perfect. It won't catch cases where there is a small prefix that should have been longer, for example.
A better setup would be to have profiling enabled on all queries and logging the sum of all predicates looked into. Then we could look at queries that took long, traversed many facts and return few results for a more accurate identification of problematic queries.
But enabling profiling all the time will have performance cost to the queries, whilst identifying full scans is basically free.

Note: The part of actually logging the data is not in this diff. I'll do that if we decide to go ahead.

Reviewed By: simonmar

Differential Revision: D47224076

fbshipit-source-id: eb2a4d0f1c8389718befe094b1ef77fa04181bd2