[oss] Use HTTP transport, drop fbthrift dependency
Changes earlier removed the dependency on fbthrift serialization in
the C++ parts of Glean (mainly glean/rts). The only remaining
dependency on fbthrift is for the transport layer between server and
client.

I've added a simple HTTP transport layer for Thrift (see
https://github.com/facebookincubator/hsthrift/pull/111). This PR moves
the OSS build of Glean to use the HTTP transport, and drops the
fbthrift dependency by default. It's still available via flags if we
need it in the future.