Replace `predicate Uses` with `type From`.
Summary:
The attempt to share `Uses` by making it a predicate didn't work out so well for the derive passes. The time it took to retrieve all of the nested facts within all of the `FileXRefMap`s seems like it was too much.

This diff re-introduces `Uses` but this time as a `type`, named `From`.

Reviewed By: lazamar, simonmar

Differential Revision: D47251705

fbshipit-source-id: 747bb529d7befb85a10699d573dd6f9e127da763