Fix up unresolved hack.FileCalls that are actually resolved
Summary:
I discovered the vast majority of calls to
DetectionZonedValue::genUnwrap were captured as codemarkup.EntityUses but only as
occurrences (unresolved entity references) in hack.FileCall, and
indeed that there are duplicate resolved and unresolved entity
references with the same span data for each of them.  This works
around the problem for the moment.

Reviewed By: pepeiborra

Differential Revision: D47296168

fbshipit-source-id: 65c2f31785e23c7fb989d9317d17b81c7c7ac6e1