Sketcher cleanup: Put definition state into a monad
Summary:
The eventual goal here is to structure things so deeply nested
definitions pull their references out of the reference set early, and
classes pull their methods out of the function set early, and then
top-level functions are what remain and we can address some of the
quadratic behavior of reference matching.  But for the moment this is
very simple and just hides state away in a monad transformer but is
otherwise behaviorally unchanged.

Reviewed By: pepeiborra

Differential Revision: D47533644

fbshipit-source-id: 3ab9554c99a2c1d5ef34b995cf6f27743f1f0446