override repo mapping with provided db
Summary:
When computing a glass snapshot, we use a local db. In order to conform with  Glass hard-coded repo mapping, we need to give this db the expected name.

https://www.internalfb.com/code/fbsource/[6e249b30868a]/tools/skycastle/lib2/glean/diff_sketch/core.sky?lines=561

This isn't ideal.  We want to make the snapshot invocation language agnostic, and possibly use the same db for several languages.

This diff adds a new option to the snapshot tool to specify the DB to use.

```
> buck2 run //glean/glass:glass-snapshot -- --help
...
  --db NAME/INSTANCE       identifies the database
```

Reviewed By: pepeiborra

Differential Revision: D47614599

fbshipit-source-id: 98294704a15f7fe58affcf78c43fedc697165b96