Report a single, name-preferred, range for symbols.
Summary:
This is step 2 in this proposed plan starting in D47707335.

It also finishes the job for "Navigation to an entity within a macro definition doesn't work" section of [Improving Macro Handling for Glean](https://fb.workplace.com/notes/1400996680749241), and supercedes D46067710.

We follow the xref jump to its destination, which may be in a different file.

## Proposed Plan

Currently, the full range of the symbols are used as backup by clients in the cases where the name range doesn't exist, but this fallback logic can simply happen in Glass instead.

This simplifies the Glass thrift schema, and transfers less data.

Rather than adding yet another field and trying to do this migration, I think we can just re-use the existing `target` and `range` fields to do all that we want.

1. Back out the `destination` field. I believe it's soon enough that no one is using this yet.
2. **Make Glass prioritize the name range in the existing `target` and `range` fields. The clients currently manually perform the name range prioritization, so this will be a no-op for them.**
3. Update the clients to just access the `target` and `range` fields
4. Remove the name range Thrift fields.

Reviewed By: donsbot

Differential Revision: D47698435

fbshipit-source-id: 59131c7d8286c31ebade9eae56abc207319d280e