Change indexer to emit src.ByteSpan for method locations
Summary:
the Kotlin MethodDeclaration was using src.Loc for locations. This is not usable by most things as its only a point (column/line) in the file. We need spans. Switch to the more concise and easier to generate src.ByteSpan.

We use a `maybe` to allow the new field to be backwards compatible.

- update schema
- update indexer
- regenerate tests

To generate the new field we need to bump the indexer via its dotslash: here : https://www.internalfb.com/intern/msdk/bump/?action=view&schedule_fbid=789173989387026

Reviewed By: lazamar

Differential Revision: D47583757

fbshipit-source-id: bcb94dbd25947b85bf0244f9c1fee9cfc6c1fd6d