Account for key clashes in FactSet rebase
Summary:
This diff addresses one of the problems listed in https://fb.workplace.com/groups/csi.wtf/posts/6416068098469802/

After a FactSet rebase we may end-up with two facts which after substitutions have the same key.
This is perfectly legal and we should allow that.

We may also end-up with two facts that after substitution have the same key, but different values.
Here we will also allow that and ignore the second key definition with the conflicting fact.
We do that because these can arise from the fact that the value in one of these could refer to a remote fact id while the other one refers to a local fact id.
These may or may not ultimately map to the same thing, but in the client our cache already evicted the facts we needed to determine that.To avoid failing on legitimate facts we will make a concession and allow this.

Reviewed By: zsol

Differential Revision: D47725527

fbshipit-source-id: 2556a0669720c4cab1b153a6fc0a95743153b49a