Allow value conflicts in SendAndRebaseQueue FactSet
Summary:
see https://fb.workplace.com/groups/csi.wtf/permalink/6416068098469802/

Redefinition is when we have two facts A->B, C->D where A == C but B /= D.

Normally this would be an error, and defineBatch rejects it with an
error. But it can arise naturally here because when
typechecking/rebasing A->B we may have used a different cache from
when we typechecked/rebased C->D. So even though B and D may be
semantically identical, they are literally different.

It doesn't seem possible to ensure that we always use a consistent
cache, so it's inevitable that benign fact redefinitions may
occur. Therefore we ignore redefinitions in `defineBatch`.

Note: this might mean that we are ignoring actual errors and silently
picking one of the two facts if they really did differ. That's bad,
but I don't see an alternative.

Reviewed By: zsol

Differential Revision: D47672960

fbshipit-source-id: 453d942913d63c2b3020e9d3b5f8f4406c361d92