Remove the name range fields.
Summary:
This is step 4 in this proposed plan starting in D47707335.

## Proposed Plan

Currently, the full range of the symbols are used as backup by clients in the cases where the name range doesn't exist, but this fallback logic can simply happen in Glass instead.

This simplifies the Glass thrift schema, and transfers less data.

Rather than adding yet another field and trying to do this migration, I think we can just re-use the existing `target` and `range` fields to do all that we want.

1. Back out the `destination` field. I believe it's soon enough that no one is using this yet.
2. Make Glass prioritize the name range in the existing `target` and `range` fields. The clients currently manually perform the name range prioritization, so this will be a no-op for them.
3. Update the clients to just access the `target` and `range` fields
4. **Remove the name range Thrift fields.**

Reviewed By: donsbot

Differential Revision: D47709052

fbshipit-source-id: 71c3cdf60af340cd9fc79ab4adb2eb0f83cd0198