Introduce 'recursive' construct
Summary:
This diff adds the bytecode of the `rec` construct from the recursion docs.

With this we will have a naive evaluation. i.e. run the entire query multiple times until we stop finding new facts.

In the next steps I will introduce the code transformation to make this into a semi-naive evaluation but that will be done in the IR, not in the bytecode.

Currently the construct is checking the id bounds of the underlying `Define`.
In the future, to improve the performance of adding and seeking ad-hoc recursive events this will check the bounds of a dedicated `Define` used by the `recursive` constructs.

Although the particular example in the test works, this doesn't make everything work yet. We still don't have a way to assign different "arguments" to different evaluations of a recursive predicate. This is coming.

For background see [Evaluating recursive queries](https://fb.quip.com/mj5MAfYdlD0a)

Reviewed By: nhawkes

Differential Revision: D47669357

fbshipit-source-id: d298efdb28343aaeb4f4d055735312a964c77663