Stop indexing classes/interfaces/enums multiple times
Summary:
By recurring in the visitor, and recurring in member elements, nested classes are being visited
multiple times. There used to be a check for this in a hashtable, but that failed due to duplication
in keys from not actually converting the keys properly.

This will omit any classes nested in anonymous classes, but those wouldn't have been covered properly now anyway. Need a new visitor.

Reviewed By: ermattt, donsbot

Differential Revision: D47917593

fbshipit-source-id: 59b9c81298245e5f20988b0b6ef3a6580cb839f1