Return decls in cases where there is no definition representative
Summary:
For e.g. thrift service definitions, there's only a declaration token. These are still really useful to search for
We've been filtering them out, but it was a bit too strict. Return the decl entity in the case the definition isn't there.

An idea discussed with Michael Park to improve perf would be to select a representiative of the decl family (in the index) as the search result. We can't do that sort of selection yet.

Reviewed By: phlalx

Differential Revision: D48143928

fbshipit-source-id: 82325b676e1b63509a26c77ef2987ccc1e01b286